var-wireless-utils: disable wpa_supplicant@wlan0.service by default
On some platforms, such as i.MX93-based SOMs, the wlan0 interface may be
absent due to configuration choices—for example, when the CM33 device tree
is used. In these cases, enabling the wpa_supplicant@wlan0.service causes
boot-time delays as systemd waits for the non-existent device:

  ...
  [ ***  ] A start job is running for /sys/sub…devices/wlan0 (1min 3s / 1min 30s)
  ...
  [ TIME ] Timed out waiting for device /sys/subsystem/net/devices/wlan0.
  [DEPEND] Dependency failed for WPA supplica…aemon (interface-specific version).

The `wpa_supplicant@wlan0.service` is in charge for managing the `wpa_supplicant`
daemon for the `wlan0` interface, enabling Wi-Fi connection. It can still
be enabled manually as part of Wi-Fi setup.

This commit disables the service by default to avoid such delays and
dependency failures on configurations where Wi-Fi is not present.

This reverts commit a0f28423aa884aff42dd5b658098bb89acbfb28d.